Change to auto-generation
Previously, useers selected and submitted options to generate a poem.
Now, landing on the site auto-generates a poem of 10 sections,
10 lines per section, allowing repeat lines. Manually refreshing the page
generates a new poem as well. The page auto-refreshes, and a new poem
is generated, after 10 minutes, and there's a countdown in the top-right
to show the remaining time until that happens. Each time, the poem
is saved to the db.

This PR also adds two new pages:

- A page that shows the generation timestamp and first line of the 20
most recent poems. Clicking a first line then retrieves and displays
the full poem.
- A page that provides some context for the site.

The previous version of this app is saved in rotations_user_input.